+++ HISTORY.md
@@ -3,6 +3,7 @@
 
   * Remove private `req.session.save()`
   * Remove undocumented `req.session.length` to free up key name
+  * Remove undocumented `req.sessionCookies` and `req.sessionKey`
   * deps: cookies@0.8.0
     - Fix check for default `secure` option behavior
     - Fix `maxAge` option preventing cookie deletion